refactor: split game component
