+++ src/components/Sprint-game/finish-page/finish-page.module.css
@@ -1,6 +1,6 @@
 .finish-game-wrapper {
   width: 430px;
-  min-height: 390px;
+  /* min-height: 390px; */
   height: 100%;
   background-color: #ffeeff;
   border-radius: 30px;
