+++ src/components/Sprint-game/game-page/game-page.module.css
@@ -1,0 +1,183 @@
+.game__wrapper.play {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  justify-content: space-between;
+  box-sizing: border-box;
+  /* padding: 30px; */
+  min-height: 570px;
+  min-width: 1022px;
+  /* margin-top: auto; */
+}
+.game__field {
+  max-width: 710px;
+  width: 100%;
+  min-height: 460px;
+  /* height: 100%; */
+  max-height: 460px;
+  background-color: #ffff;
+  border-radius: 30px;
+  margin: 0 1.5rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 2rem 2rem 3.5rem;
+  margin-top: 35px;
+}
+.cat__img1 {
+  position: absolute;
+  right: -100px;
+}
+.side__buttons {
+  width: 120px;
+  display: flex;
+  justify-content: space-between;
+}
+img.side__buttons {
+  padding: 1rem;
+}
+
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  /* margin-right: 40px; */
+  margin: 40px;
+}
+.watch__wrapper {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 1.8rem;
+  margin-top: 3rem;
+}
+
+.watch__wrapper > svg {
+  margin-bottom: 1.5rem;
+}
+.timer2 {
+  position: absolute;
+}
+
+.guess__word {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  font-size: 2.5rem;
+}
+
+.green__button,
+.red__button {
+  width: 20rem;
+  height: 4rem;
+  border-radius: 2rem;
+  border: none;
+
+  color: #733999;
+  font-size: 1.8rem;
+  font-weight: 900;
+  cursor: pointer;
+}
+.green__button {
+  margin-right: 8rem;
+  background-color: #45f9a3;
+}
+
+.red__button {
+  background-color: #f945d1;
+}
+
+.button__toguess {
+  display: flex;
+  margin: 0 auto;
+}
+
+.points {
+  display: flex;
+  flex-direction: column;
+}
+.total__points {
+  font-size: 2.4rem;
+  font-weight: 900;
+  color: #45f9a3;
+}
+.current__points {
+  color: #ffff;
+  font-weight: 900;
+  font-size: 2.4rem;
+}
+
+.point-number {
+  margin-right: 0.5rem;
+}
+
+.check__points {
+  height: 100px;
+}
+
+.guess_not {
+  display: flex;
+  justify-content: center;
+  height: 30px;
+}
+.lines {
+  height: 1px;
+  width: 480px;
+  background-color: #733999;
+  opacity: 0.6;
+  position: relative;
+}
+
+.correct__sign {
+  position: absolute;
+  bottom: 31%;
+}
+.refresh__button,
+.close__button {
+  background-color: transparent;
+  /* border-radius: 100%; */
+  border: none;
+}
+
+.full-screen-game {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #fdebff;
+  border-radius: 50px;
+  max-width: 1022px;
+  width: 80%;
+  min-height: 560px;
+  text-align: center;
+  padding: 5px;
+}
+
+.full-screen__button {
+  position: absolute;
+  right: 66px;
+  top: 20px;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  height: 30px;
+  width: 30px;
+
+  background-color: rgb(115, 57, 153);
+  border-radius: 50%;
+  border: none;
+  cursor: pointer;
+
+  color: white;
+}
+svg.full-screen__icon {
+  width: 24px;
+  height: 24px;
+}
