+++ src/components/Sprint-game/sprint-game/SprintGame.tsx
@@ -1,36 +1,17 @@
-import { request } from 'node:http';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import useSound from 'use-sound';
 import {
   clearWords,
   sprintGameBallsCounter,
   sprintGameCheckPoints,
-  sprintGameCurrentPoints,
-  sprintGameRandomArray,
-  sprintGameSetLearntWords,
-  sprintGameSetNotLearntWords,
-  sprintGameShuffledArray,
   sprintGameStatusChange,
-  sprintGameTotalPoints,
+  sprintGameTotalPoints
 } from '../../../actions/sprint-game-action';
-import { fetchWordsList } from '../../../actions/word-actions';
 import { ReactComponent as Cat } from '../../../assets/images/cat2.svg';
-import closeIcon from '../../../assets/images/close.svg';
-import correctImage from '../../../assets/images/correct.svg';
-import inCorrectImage from '../../../assets/images/incorrect.svg';
-import questionIcon from '../../../assets/images/question.svg';
-import refreshIcon from '../../../assets/images/refreshing.svg';
-import banner from '../../../assets/images/sprint-top.png';
 import { ReactComponent as Timer1 } from '../../../assets/images/timer1.svg';
-import { ReactComponent as Timer2 } from '../../../assets/images/timer2.svg';
-import countDown from '../../../assets/sounds/countDown.wav';
-import correctSound from '../../../assets/sounds/src_music_correct.mp3';
-import wrongSound from '../../../assets/sounds/src_music_wrong.wav';
-import { RootStateType } from '../../../reducer/root-reducer';
-import Balls from '../balls/Balls';
-import CheckPoints from '../check-points/CheckPoints';
+import GamePage from '../game-page/GamePage';
 import FinishPage from '../finish-page/FinishPage';
+import { RootStateType } from '../../../reducer/root-reducer';
 import { Timer } from '../timer/Timer';
 import { TitleGamePage } from '../title-page/TitleGamePage';
 import styles from './sprint-game.module.css';
@@ -38,9 +19,6 @@ import styles from './sprint-game.module.css';
 const SprintGame: React.FC = () => {
   const dispatch = useDispatch();
 
-  const wordList = useSelector(
-    (state: RootStateType) => state.wordState.currentWordList
-  );
   const gameStatuses = useSelector(
     (state: RootStateType) => state.sprintGameState
   );
@@ -49,51 +27,9 @@ const SprintGame: React.FC = () => {
     gameTitle,
     gameDescription,
     gameStatus,
-    totalPoints,
-    currentPoints,
-    shuffledArray,
-    randomArray,
-    checkpoints,
-    ballsCounter,
   } = gameStatuses;
 
-  const [wordCounter, setWordCounter] = useState(0);
-  const [wordToGuess, setWordToGuess] = useState('');
-  const [correctAnswer, setCorrectAnswer] = useState(true);
-  const [playCorrectSound] = useSound(correctSound, {
-    interrupt: true,
-  });
-  const [playWrongSound] = useSound(wrongSound, { interrupt: true });
-  const getRandomNumber = (num: number) => Math.floor(Math.random() * num);
 
-  useEffect(() => {
-    if (wordList.length === 0) {
-      dispatch(fetchWordsList({ page: 0, group: 0 }));
-    }
-    dispatch(sprintGameRandomArray(createRandomArray()));
-  }, []);
-
-  useEffect(() => {
-    dispatch(sprintGameStatusChange('start'));
-    dispatch(clearWords());
-    dispatch(sprintGameTotalPoints(0));
-    dispatch(sprintGameBallsCounter(0));
-    dispatch(sprintGameCheckPoints(0));
-  }, []);
-
-  const createRandomArray = () => {
-    const array = [];
-    for (let i = 0; i < wordList.length; i++) {
-      array.push(getRandomNumber(wordList.length));
-    }
-    return array;
-  };
-
-  useEffect(() => {
-    dispatch(
-      sprintGameShuffledArray(wordList.slice().sort(() => Math.random() - 0.5))
-    );
-  }, [wordList]);
 
   const renderTimerPage = () => (
     <div className={`${styles.game__wrapper} ${styles.timer__page}`}>
@@ -106,168 +42,13 @@ const SprintGame: React.FC = () => {
     </div>
   );
 
-  const changeGameStats = () => {
-    dispatch(sprintGameSetLearntWords(shuffledArray[wordCounter]));
-    changeCurrentPoints();
-    setCorrectAnswer(true);
-    playCorrectSound();
-    dispatch(sprintGameTotalPoints(totalPoints + currentPoints));
-    dispatch(sprintGameCheckPoints(checkpoints < 3 ? checkpoints + 1 : 1));
-    // checkTheEndOfTheGame();
-    if (checkpoints === 2) {
-      dispatch(sprintGameBallsCounter(ballsCounter + 1));
-      if (ballsCounter === 4) {
-        dispatch(sprintGameBallsCounter(ballsCounter));
-      }
-    }
-  };
-
-  const cleanCurrentGameStats = () => {
-    dispatch(sprintGameSetNotLearntWords(shuffledArray[wordCounter]));
-    // checkTheEndOfTheGame();
-    setCorrectAnswer(false);
-    playWrongSound();
-    dispatch(sprintGameCheckPoints(0));
-  };
-
-  const checkTheWordRight = () => {
-    if (shuffledArray[wordCounter].wordTranslate === wordToGuess) {
-      changeGameStats();
-    } else {
-      cleanCurrentGameStats();
-    }
-    checkTheEndOfTheGame();
-    setWordCounter(wordCounter + 1);
-  };
-
-  const checkTheWordWrong = () => {
-    if (shuffledArray[wordCounter].wordTranslate !== wordToGuess) {
-      changeGameStats();
-    } else {
-      cleanCurrentGameStats();
-    }
-    checkTheEndOfTheGame();
-    setWordCounter(wordCounter + 1);
-  };
-
-  const changeCurrentPoints = () => {
-    if (ballsCounter === 0) {
-      dispatch(sprintGameCurrentPoints(50));
-    } else if (ballsCounter === 1) {
-      dispatch(sprintGameCurrentPoints(60));
-    } else if (ballsCounter === 2) {
-      dispatch(sprintGameCurrentPoints(70));
-    } else if (ballsCounter === 3) {
-      dispatch(sprintGameCurrentPoints(80));
-    } else if (ballsCounter === 4) {
-      dispatch(sprintGameCurrentPoints(100));
-    }
-  };
-
   useEffect(() => {
-    console.log(wordCounter, 'error here');
-    console.log(shuffledArray, 'array');
-    console.log();
-    if (wordCounter < shuffledArray.length) {
-      getRandomNumber(2) === 0
-        ? setWordToGuess(shuffledArray[wordCounter].wordTranslate)
-        : setWordToGuess(shuffledArray[randomArray[wordCounter]].wordTranslate);
-    }
-  }, [wordCounter]);
-
-  const checkTheEndOfTheGame = () => {
-    if (wordCounter === shuffledArray.length - 1) {
-      dispatch(sprintGameStatusChange('finish'));
-    }
-  };
-
-  const renderGamePage = () => (
-    <div className={`${styles.game__wrapper} ${styles.play}`}>
-      <div className={styles.sidebar}>
-        <div className={styles.watch__wrapper}>
-          <Timer initialTimer={60} nextPage="finish" timerFontSize="1.8rem" />
-          <Timer2 className={styles.timer2} />
-        </div>
-      </div>
-      <div
-        className={styles.game__field}
-        style={{
-          backgroundImage: `url(${banner})`,
-          backgroundRepeat: 'no-repeat',
-        }}
-      >
-        <div className={styles.point}>
-          <div className={styles.total__points}>{totalPoints}</div>
-          <div className={styles.current__points}>
-            <span className={styles['point-number']}>
-              {currentPoints > 0 ? currentPoints : currentPoints}
-            </span>
-            очков за слово
-          </div>
-        </div>
-
-        <div className={styles.check__points}>
-          <CheckPoints />
-          <Balls />
-        </div>
-        <div className={styles.guess__word}>
-          <div className={styles.the__word}>
-            {shuffledArray[wordCounter].word}
-          </div>{' '}
-          -<div className={styles.translation}>{wordToGuess}</div>
-        </div>
-
-        <div className={styles.guess_not}>
-          <div className={styles.lines}> </div>
-          <img
-            src={correctAnswer ? correctImage : inCorrectImage}
-            alt="guessed-or-not"
-            className={styles.correct__sign}
-          />
-        </div>
-        <div className={styles.button__toguess}>
-          <button
-            type="button"
-            className={styles.green__button}
-            onClick={checkTheWordRight}
-          >
-            Верно
-          </button>
-          <button
-            type="button"
-            className={styles.red__button}
-            onClick={checkTheWordWrong}
-          >
-            Неверно
-          </button>
-        </div>
-      </div>
-
-      <div className={styles.side__buttons}>
-        <button
-          type="button"
-          className={styles.refresh__button}
-          onClick={() => dispatch(sprintGameStatusChange('timer '))}
-        >
-          <img src={refreshIcon} alt="refresh icon" />
-        </button>
-        <button
-          type="button"
-          className={styles.question__button}
-          onClick={() => console.log('question?')}
-        >
-          <img src={questionIcon} alt="question icon" />
-        </button>
-        <button
-          type="button"
-          className={styles.close__button}
-          onClick={() => dispatch(sprintGameStatusChange('start'))}
-        >
-          <img src={closeIcon} alt="close icon" />
-        </button>
-      </div>
-    </div>
-  );
+    dispatch(sprintGameStatusChange('start'));
+    dispatch(clearWords());
+    dispatch(sprintGameTotalPoints(0));
+    dispatch(sprintGameBallsCounter(0));
+    dispatch(sprintGameCheckPoints(0));
+  }, []);
 
   return (
     <div className={styles.sprint__game}>
@@ -279,7 +60,7 @@ const SprintGame: React.FC = () => {
         />
       ) : null}
       {gameStatus === 'timer' ? <div>{renderTimerPage()}</div> : null}
-      {gameStatus === 'play' ? <div>{renderGamePage()}</div> : null}
+      {gameStatus === 'play' ? <GamePage /> : null}
       {gameStatus === 'finish' ? <FinishPage /> : null}
     </div>
   );
